+++ src/EncompassRest.Loans.v3/Loans/v3/Employment.cs
@@ -220,9 +220,9 @@ public sealed partial class Employment : DirtyExtensibleObject, IIdentifiable
     public decimal? MilitaryPropPay { get => GetValue<decimal?>(); set => SetValue(value); }
 
     /// <summary>
-    /// Employment MonthlyIncomeAmount
+    /// Employment MonthlyIncomeAmount [FE0112]
     /// </summary>
-    public int? MonthlyIncomeAmount { get => GetValue<int?>(); set => SetValue(value); }
+    public decimal? MonthlyIncomeAmount { get => GetValue<decimal?>(); set => SetValue(value); }
 
     /// <summary>
     /// Employment NoLinkToDocTrackIndicator [BE0097], [CE0097]
